[autostart] moved out of utils
